+++ src/main/java/nu/olivertwistor/currencymgr/database/package-info.java
@@ -1,1 +1,7 @@
+/**
+ * This package contains code for the database, connections and things like
+ * that.
+ *
+ * @since 0.1.0
+ */
 package nu.olivertwistor.currencymgr.database;