+++ src/main/java/nu/olivertwistor/currencymgr/ui/actions/AbstractDialogSaveAction.java
@@ -1,6 +1,7 @@
-package nu.olivertwistor.currencymgr.ui;
+package nu.olivertwistor.currencymgr.ui.actions;
 
 import nu.olivertwistor.currencymgr.database.CurrencyFile;
+import nu.olivertwistor.currencymgr.ui.GUI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -8,22 +9,32 @@ import javax.swing.AbstractAction;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 import java.awt.event.ActionEvent;
-import java.awt.event.KeyEvent;
 import java.io.File;
 import java.sql.SQLException;
 
-public final class NewFileAction extends AbstractAction
+/**
+ * This abstract class provides a common {@link #actionPerformed(ActionEvent)}
+ * where a save dialog opens. It doesn't set any of the usual
+ * {@link AbstractAction} values, such as name, description, icon and mnemonic.
+ *
+ * @since 0.1.0
+ */
+@SuppressWarnings("HardCodedStringLiteral")
+abstract class AbstractDialogSaveAction extends AbstractAction
 {
     private static final Logger LOG = LogManager.getLogger();
 
     private final GUI gui;
 
-    public NewFileAction(final GUI gui)
+    /**
+     * Creates a new abstract dialog save action.
+     *
+     * @param gui the gui of which the save dialog is a parent
+     *
+     * @since 0.1.0
+     */
+    AbstractDialogSaveAction(final GUI gui)
     {
-        super("New...");
-        this.putValue(SHORT_DESCRIPTION, "Create a new file.");
-        this.putValue(MNEMONIC_KEY, KeyEvent.VK_N);
-
         this.gui = gui;
     }
 
@@ -31,9 +42,8 @@ public final class NewFileAction extends AbstractAction
     public void actionPerformed(final ActionEvent e)
     {
         final JFileChooser fileChooser = new JFileChooser();
-        fileChooser.addChoosableFileFilter(new CurrencyFileFilter());
         fileChooser.setAcceptAllFileFilterUsed(true);
-        fileChooser.setDialogTitle("Create a new currency manager file");
+        fileChooser.setDialogTitle("Save currency file");
 
         final int status = fileChooser.showSaveDialog(this.gui);
         if (status == JFileChooser.APPROVE_OPTION)
@@ -51,12 +61,21 @@ public final class NewFileAction extends AbstractAction
             {
                 JOptionPane.showMessageDialog(
                         this.gui,
-                        "The currency manager file couldn't be saved.",
+                        "The currency file couldn't be saved.",
                         "Failed to save file",
                         JOptionPane.ERROR_MESSAGE,
                         null);
-                LOG.fatal("Failed to save currency manager file.", exception);
+                LOG.fatal("Failed to save currency file.", exception);
             }
         }
+    }
+
+    @SuppressWarnings("PublicMethodWithoutLogging")
+    @Override
+    public String toString()
+    {
+        return "AbstractDialogSaveAction{" +
+                "gui=" + this.gui +
+                '}';
     }
 }