+++ src/main/java/nu/olivertwistor/currencymgr/ui/package-info.java
@@ -1,1 +1,6 @@
+/**
+ * This package contains code for the UI and also code supporting the UI.
+ *
+ * @since 0.1.0
+ */
 package nu.olivertwistor.currencymgr.ui;