+++ CHANGELOG.md
@@ -3,5 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][1] and this project adheres to [Semantic Versioning 2.0.0][2].
 
+## Unreleased
+
+## [0.1.0](https://github.com/olivertwistor/currency-manager/releases/v0.1.0)
+* Added
+    * GUI main window
+    * Ability to create a new currency file
+    * Ability to open an existing currency file
+
 [1]: https://keepachangelog.com/en/1.0.0/
 [2]: https://semver.org/